worker/requirements.txt
confluent-kafka==2.3.0
docker==7.1.0
psycopg2-binary==2.9.9
redis==5.0.1
structlog==23.2.0
python-dotenv==1.0.0
